modify .prettierrc.json/pakage.json/.eslintrc.json  and update README.md
